Remove vscode dependency from bazel_command
bazel_command currently has a dependency on vscode for the Bazel executable path. This change is to remove that dependency by passing the executable path into the constructor of the BazelCommand class.

This has the following benefits:
1. This will allow us to eventually move the Bazel-specific logic into a separate library.
2. It makes the code more testable.
