+++ src/bazel/bazel_quickpick.ts
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 import * as vscode from "vscode";
+import { getDefaultBazelExecutablePath } from "../extension/configuration";
 import { IBazelCommandAdapter, IBazelCommandOptions } from "./bazel_command";
 import { BazelQuery } from "./bazel_query";
 import { BazelWorkspaceInfo } from "./bazel_workspace_info";
@@ -73,6 +74,7 @@ async function queryWorkspaceQuickPickTargets(
   query: string,
 ): Promise<BazelTargetQuickPick[]> {
   const queryResult = await new BazelQuery(
+    getDefaultBazelExecutablePath(),
     workspaceInfo.bazelWorkspacePath,
     query,
     [],
@@ -96,6 +98,7 @@ async function queryWorkspaceQuickPickPackages(
   workspaceInfo: BazelWorkspaceInfo,
 ): Promise<BazelTargetQuickPick[]> {
   const packagePaths = await new BazelQuery(
+    getDefaultBazelExecutablePath(),
     workspaceInfo.bazelWorkspacePath,
     "...:*",
     [],
