+++ src/symbols/bazel_target_symbol_provider.ts
@@ -20,6 +20,7 @@ import {
   getTargetsForBuildFile,
   QueryLocation,
 } from "../bazel";
+import { getDefaultBazelExecutablePath } from "../extension/configuration";
 import { blaze_query } from "../protos";
 
 /** Provids Symbols for targets in Bazel BUILD files. */
@@ -38,6 +39,7 @@ export class BazelTargetSymbolProvider implements DocumentSymbolProvider {
     }
 
     const queryResult = await getTargetsForBuildFile(
+      getDefaultBazelExecutablePath(),
       workspaceInfo.bazelWorkspacePath,
       document.uri.fsPath,
     );
