+++ src/workspace-tree/bazel_package_tree_item.ts
@@ -19,6 +19,7 @@ import {
   IBazelCommandAdapter,
   IBazelCommandOptions,
 } from "../bazel";
+import { getDefaultBazelExecutablePath } from "../extension/configuration";
 import { blaze_query } from "../protos";
 import { BazelTargetTreeItem } from "./bazel_target_tree_item";
 import { IBazelTreeItem } from "./bazel_tree_item";
@@ -53,6 +54,7 @@ export class BazelPackageTreeItem
 
   public async getChildren(): Promise<IBazelTreeItem[]> {
     const queryResult = await new BazelQuery(
+      getDefaultBazelExecutablePath(),
       this.workspaceInfo.bazelWorkspacePath,
       `//${this.packagePath}:all`,
       [],
